Rewritten whole project from Commonjs to ES modules.
Svelte will only support ES modules, so all the project had to be moved from Commonjs to ES modules.

There's a new bug in pdf-parse which since all the project is called as ES modules and it's written in Commonjs, turns on the Debug mode and throws unknown errors.
It was bypassed by modifying the pdf-parse/index.js inside node_modules to disable it permanently.